Register: Convert int to long, finish refactor

+++ src/main/java/CashRegisterModel.java
@@ -7,7 +7,7 @@ import main.java.resources.RegisterContents;
 import java.util.Map;
 
 public class CashRegisterModel {
-    private Map<String, Integer> contents;
+    private Map<String, Long> contents;
     private static DollarValueConstants dollars;
 
     public CashRegisterModel() {
@@ -15,12 +15,12 @@ public class CashRegisterModel {
         dollars = new DollarValueConstants();
     }
 
-    public Map<String, Integer> getContents() {
+    public Map<String, Long> getContents() {
         return contents;
     }
 
-    public int getBalance() {
-        int value = 0;
+    public long getBalance() {
+        long value = 0;
 
         for(String key: contents.keySet()) {
             value += computeDenominationValue(key);
@@ -29,15 +29,15 @@ public class CashRegisterModel {
         return value;
     }
 
-    private int computeDenominationValue(String bill) {
+    private long computeDenominationValue(String bill) {
         return contents.get(bill) * dollars.get(bill);
     }
 
-    public void deposit(Map<String, Integer> deposit) {
-        deposit.forEach((key, value) -> contents.merge(key, value, Integer::sum));
+    public void deposit(Map<String, Long> deposit) {
+        deposit.forEach((key, value) -> contents.merge(key, value, Long::sum));
     }
 
-    public void withdraw(Map<String, Integer> withdrawal) throws InsufficientFundsException {
+    public void withdraw(Map<String, Long> withdrawal) throws InsufficientFundsException {
         boolean insufficient = contents.keySet().stream().anyMatch((key) -> withdrawal.get(key) > contents.get(key));
 
         if (insufficient) {
